更新：给网关下发配置参数

+++ src/main/java/com/example/qmx/domain/Sensor.java
@@ -19,7 +19,7 @@ public class Sensor {
     private String devName;
 
     @TableField("value")
-    private Double value;
+    private Float value;
 
     @TableField("time")
     private Date time;
